c3ms-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
openai